get-current-user-level util
